Fixed bug where turrets didn't update new target when its previous target died.
Added StoreEntityParentComponent.java to make entities able to store the entity who spawned it.

Changed how the turrets and enemies list was retrieved. Now only the turret which was added is added to the list instead of adding the entire list again. The same for enemies, which also gets removed when a collision is detected.